+++ packages/client/src/features/repo-search/model/index.ts
@@ -1,1 +1,2 @@
 export { useRepoUrlForm } from "./useRepoUrlForm";
+export { useCommits } from "./useCommits";
