--- packages/client/src/shared/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useCommits } from "./useCommits";
